Welp, it seems to work... Added test for compression too. LoadFile should return a void pointer dang it

+++ plcore/include/plcore/pl_compression.h
@@ -22,8 +22,8 @@ typedef enum PLCompressionType {
 PL_EXTERN_C
 
 void *PlCompress_Deflate( const void *src, unsigned long srcLength, unsigned long *dstLength );
-void *PlCompress_LZRW1( const void *src, unsigned long length, unsigned long *outLength );
+void *PlCompress_LZRW1( const void *src, size_t length, size_t *outLength );
 
-void *PlDecompress_LZRW1( const void *src, unsigned long srcLength, unsigned long *dstLength );
+void *PlDecompress_LZRW1( const void *src, size_t srcLength, size_t *dstLength );
 
 PL_EXTERN_C_END
